add first component to fetch categories;
