+++ src/main/webapp/reactjs/public/edit-item.html
@@ -2,19 +2,19 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <link rel="icon" href="../assets/logo.svg">
+    <link rel="icon" href="../../assets/logo.svg">
     <title>Add Item</title>
-    <link rel="stylesheet" href="../css/add-item.css">
-    <link rel="stylesheet" href="../css/edit-item.css">
-    <link rel="stylesheet" href="../css/header-and-body.css">
+    <link rel="stylesheet" href="../../css/add-item.css">
+    <link rel="stylesheet" href="../../css/edit-item.css">
+    <link rel="stylesheet" href="../../css/header-and-body.css">
 </head>
 <body>
     <div class="header">
-        <a href="index.html"><img id="logo" src="../assets/logo.svg"></a>
+        <a href="../../html/index.html"><img id="logo" src="../../assets/logo.svg"></a>
         <div class="menu">
             <ul>
-                <li><a href="about.html">Help</a></li>
-                <li><a href="sign-up.html">Michał Kuk</a></li>
+                <li><a href="../../html/about.html">Help</a></li>
+                <li><a href="../../html/sign-up.html">Michał Kuk</a></li>
             </ul>
         </div>
     </div>
@@ -28,14 +28,7 @@
             <h2>Item Info</h2>
             <h4><label for="quest-type">Category</label><br></h4>
             <select class= "type-selector" id="quest-type" name="quest-type">
-                <option value="" selected disabled hidden>Current item category</option>
-                <option>Home equipment</option>
-                <option>Electronic</option>
-                <option>Cameras</option>
-                <option>Musical instruments</option>
-                <option>Electronic scooters</option>
-                <option>Adventure</option>
-                <option>Extra</option>
+
             </select>
 
             <h4>Description:</h4>
@@ -56,25 +49,25 @@
             <h4>Manage item's images</h4>
             <div class="input-image-grid">
                 <div class="item-photo-container" type="button" href="#">
-                    <img class="item-photo" src="../assets/plane.jpg">
+                    <img class="item-photo" src="../../assets/plane.jpg">
                 </div>
                 <div class="item-photo-container">
-                    <img class="item-photo 2" src="../assets/plane.jpg">
+                    <img class="item-photo 2" src="../../assets/plane.jpg">
                 </div>
                 <div class="item-photo-container">
-                    <img class="item-photo 3" src="../assets/plane.jpg">
+                    <img class="item-photo 3" src="../../assets/plane.jpg">
                 </div>
                 <div class="item-photo-container 4">
-                    <img class="item-photo" src="../assets/plane.jpg">
+                    <img class="item-photo" src="../../assets/plane.jpg">
                 </div>
                 <div class="item-photo-container 5">
-                    <img class="item-photo" src="../assets/plane.jpg">
+                    <img class="item-photo" src="../../assets/plane.jpg">
                 </div>
                 <div class="item-photo-container 6">
-                    <img class="item-photo" src="../assets/plane.jpg">
+                    <img class="item-photo" src="../../assets/plane.jpg">
                 </div>
                 <div class="add-item-photo-container 7">
-                    <a href="index.html"><img class="item-photo" style="max-width: 70px;" src="../assets/icon-with-mountain.svg"></a><br>
+                    <a href="../../html/index.html"><img class="item-photo" style="max-width: 70px;" src="../../assets/icon-with-mountain.svg"></a><br>
                 </div>
             </div>
         </div>
