+++ src/main/webapp/reactjs/public/index.html
@@ -12,8 +12,12 @@
     <title>React App</title>
   </head>
   <body>
-  <h1>TO jest ta strona html</h1>
+  <h1>To jest ta strona html na której aktualnie pracuje</h1>
+
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
+    <select class= "type-selector" id="quest-type" name="quest-type">
+
+    </select>
   </body>
 </html>
