+++ src/main/webapp/reactjs/src/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from './components/App/App';
+import FetchCategories from './components/item-details/FetchCategories';
 
 
 ReactDOM.render(
@@ -8,5 +9,12 @@ ReactDOM.render(
         <App/>
     </React.StrictMode>,
     document.getElementById('root')
+);
+
+ReactDOM.render(
+    <React.StrictMode>
+    <FetchCategories/>
+    </React.StrictMode>,
+    document.getElementById('quest-type')
 );
